Add initial script and README
This commit adds a simple python script for the object detection demo.
The script will record a video using the default camera, applies a
blue mask (extract only blue values in a certain range), and applies
the mask to the video.

This uses opencv and a tutorial can be found at:
https://docs.opencv.org/3.4/df/d9d/tutorial_py_colorspaces.html

Opencv and python must be installed. Instructions on how to do so are
in the README.md file.

No formal testing has been done since this is an early version of
a test script.