Colour detection: Rename file and add red overlay
This commit renames 'main.py' file to 'colourdet.py' to be more clear
what this file does. The file now only does colour detection for blue.
A red overlay is also added for anything in the video that is detected
as blue.

Minor testing has been done and blue objects that fit in the colourdet
range are highlited red.