Apriltag Detection: test code and image samples
A new object detection method was tested using apriltags instead of by
colour as it is more usable in different lighting when colours can vary.

This commit includes the new testing code for apriltag detection,
sample images and new comments in the README for the new library needed
(robotpy_apriltag).

Minor testing done with the samples included in the images folder and
was successful.

A similar example can be found at:
https://github.com/robotpy/examples/blob/main/AprilTagsVision/vision.py