NoIR camera: Setup and add apriltag detection
This commit sets up the NoIR camera on the Raspberry Pi and uses
it to detect apriltags.

See README.md for setup instructions and packages that are needed.

This version is able to detect the number of april tags seen.
Nothing is highlighted.

Minimal testing has been done because this is in an early stage.