Serial Communication: setup and initial code
This commit includes a detailed README for the setup required for serial
communication between the ESP32 and RPi 4b with links to useful
documents, sites, and a tutorial video.

It also includes some initial code to communicate a message between the
ESP32 and RPi 4b with polling.

Testing was performed to confirm that the UART message can be received
but testing is still required for the LED output based on the received
communication.