Integrate two cameras and uart comm
NoIR camera features are commented out because it
becomes very slow. It should be in its own thread.

This code takes the hand to be the largest contour area.
If hazard found and hand found, centroid of hazard found
and distance to hand measured. Distance sent to ESP.

This code does work but not as reliably as it should.