Camcom integration: Add multithreading
In dualcam.py, both camera values are read and interpreted
in the same while loop which is quite slow.

This commit adds a new file dualcamthread.py which splits
the noir and thermal camera capturing and processing into 2
separate threads.

Minimal testing has been done and both are set up, print,
and terminate correctly. The GUI is not working in this commit.