Serial Communication: Interrupt on Received Data
Modified the ESP c code to use interrupts. UART message is enqueued and
variables are modified based on message. Feedback loops through the
ouput sequenced determined by data until new variables indicate that it
should stop.

Code has been tested with a python script that loops through sample data
on the RPi. It cycles though a message for a distance in the further
range, then closer range then out of range. Output was as expected.