Camcom: Finish threading
Multithreading is now mostly working between NoIR
and thermal.

This has not been tested with hand and apriltag detection,
just with print statements in each thread's loop.
They are exiting correctly when CTRL+C is pressed.

The main thread might be slowing down the NoIR thread.
And when the camera display is turned on it uses 100% CPU.

So next step is to try multiprocessing instead of threading.