Camcom: add multiprocessing
Multi-threading was proving to have some performance issues,
so multi-processing was implemented instead.

This was done using python's multiprocessing library and a
very similar implementation to threads.
Data is not shared between processes, so there could be some
trouble with overlaying cameras.

Performance is appearing to be better with this implementation.
Next step is to synchronize the processes.