Camcom: Add synchronization between processes
Each process loops at different rates and therefore
the frames between cameras will capture at different times.

This commit adds synchronization with two flags 'noir_ready'
and 'thermal_ready' as well as a lock to access them.

The main thread controls when the flags are set to 1
and are ready to capture. There are manual delays in place
when doing so for now.

This has been tested by adding a print statement before each
capture with the time and they are off by approx. 0.0001 seconds.