Update main readme and software readme
The README file when first looking at the capstone page was very
outdated and is now updated with our purpose, scope, hazards, and
assumptions.

The software readme was missing a lot of info about each subfolder
and the mini release.

Both have been updated to where we are in the project.