BLE: Connection and Basic Command of LED
Developped code to create the ble server on the esp in C code as well as
a basic python script that can find and connect to the esp device and
send a command to turn on an LED through GPIO 13.

This was tested by running the python code on a computer and was
successful. Since no connections are needed between the RPi and esp, it
will also work the same and it is better than the UART communication.

This tutorial was followed for creating a ble server:
https://innovationyourself.com/esp32-bluetooth-low-energy-tutorial/

Next steps are to have a KeepAlive message that maintains and checks the
ble connection or throw an error to alert the user that connection has
stopped (this is a safety concern). I will also change the message to
take the distance and threshold parameters. I also will send back an
acknowledgement.