Add README and inter-process data transfer
Phase2 directory needed a README to outline files and next
steps in the project. That has been added with instructions
to set it up on a new pi.

This commit also cleans up the locking of noir_ready and
thermal_ready by using the non-recursive lock when initializing
the shared values.

Finally, this commit adds a manager and list which stores
hazard data (and in the future noir data) which can be
accessed between processes.

This has been tested (very basically) and thermal contour is
being updated in the main thread.