BLE: Updates to ble server and test client script
BLE server has pattern feedback, it has a constant output when it is
disconnected and then different patterns (or OFF) when sensing heat. It
takes inputs for the signal bit, the 2 thresholds, and the calculated
distance. The esp 32 is the server and the RPi is the client. The BLE
sends keep alive messages so that their connection does not time out.

Testing was completed with the RPi client test script that connects to
the esp 32 server and loops through sending different sample inputs that
would trigger the different patterns.

Next steps are to implement this into the main code.