BLE: Changes made after integrating for last demo
Fixed logic according to the numbers that the RPi camera processing is
actually sending which was just accounting for negative numbers for
distance from hazard and positive numbers for within hazard. Added more
gpios for haptics, one haptic needs to be changed because it is not
available for reading and writing.

This has been tested and is functioning but disconnects sometimes. It is
likely an issue with how the RPi is connecting and not with this esp
code because the esp server works without disconnecting when connecting
to other devices like my computer, so it can be pushed.