Split up dualcamproc and depth
Dualcamproc was becoming very clustered and a lot
of functionality was independent.

Everything has now been split into multiple files:
one for constants and shared imports, one for the
picam process, one for thermal process.
The main file is now eyecan.py.

All functionality has been split out and it is much
easier to understand.

Depth has also been added in picam. This was done by
taking the two landmarks of the hand at knuckle and
base of the wrist. Measurements were taken to determine
a slope of distance between landmarks and depth. Uses
pythagorean theorem to get 3D distance and that is sent to
the ESP.