BLE client changes
Changed how the ble client connects to be more efficient and use less whiles.
It also connects and reconnects in a separate function and waits to
reconnect before continuing.

We tested it and it was connecting consistently and staying connected.
It had some delays but they were due to errors on the ESP side.