Add depth calibration
Adds variable "calibrate_depth" that when true will ask the user
to place their hand at lowest and specified point (40cm above lowest).
From this, the depth of the hand can be calculated each time
with a unique hand size and shape.

This code is functional but depth values have not been tested with
the full setup.